Bugfix. Added unit tests for DB mapping.

+++ src/main/java/com/vi/dao/UserDAO.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface UserDAO extends EntityDAO<User> {
     User findById(long userId);
     List<User> getAll();
-    User create(String login, String password, String email, String name, Date registerDate, Date lastVisitDate, BigInteger level);
-    User update(long userId, String login, String password, String email, String name, Date registerDate, Date lastVisitDate, BigInteger level);
+    User create(String login, String password, String email, String name, Date registerDate, Date lastVisitDate, int level);
+    User update(long userId, String login, String password, String email, String name, Date registerDate, Date lastVisitDate, int level);
     void delete(long userId);
 }
